api_principal/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Flask-Cors==3.0.10
pymysql==1.0.2
requests==2.26.0
python-dotenv==0.19.0
flask-restx==0.5.1
cryptography==41.0.7
mysqlclient==2.1.1